collector/pyproject.toml
[project]
name = "punya-collector"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "datamodel-code-generator[http]>=0.35.0",
    "geoalchemy2>=0.18.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "python-daemon>=3.1.2",
    "pyzmq>=27.1.0",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.44",
]

[project.scripts]
collector = 'main:run'
sync-models = 'main:sync_models'

[dependency-groups]
dev = [
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
